tests/requirements.txt
atomicwrites==1.4.0
attrs==19.3.0
colorama==0.4.3
more-itertools==8.3.0
packaging==20.4
pluggy==0.13.1
py==1.8.1
pymongo==3.10.1
pyparsing==2.4.7
pytest==5.4.2
python-dotenv==0.13.0
selenium==3.141.0
six==1.14.0
urllib3==1.25.9
wcwidth==0.1.9
